+++ src/full/Agda/Interaction/Highlighting/Precise.hs
@@ -99,6 +99,8 @@ data OtherAspect
     -- could for instance be an emptyness constraint.
   | TerminationProblem
   | PositivityProblem
+  | ReachabilityProblem
+  | CoverageProblem
   | IncompletePattern
     -- ^ When this constructor is used it is probably a good idea to
     -- include a 'note' explaining why the pattern is incomplete.