+++ src/full/Agda/TypeChecking/Coverage.hs
@@ -25,7 +25,8 @@ import Control.Monad.Trans ( lift )
 import Control.Applicative hiding (empty)
 #endif
 
-import Data.List hiding (null)
+import Data.Either (lefts)
+import Data.List as List hiding (null)
 import Data.Monoid (Any(..))
 import Data.Map (Map)
 import qualified Data.Map as Map
@@ -66,6 +67,7 @@ import Agda.Utils.Monad
 import Agda.Utils.Null
 import Agda.Utils.Permutation
 import Agda.Utils.Size
+import Agda.Utils.Suffix (nameVariant)
 import Agda.Utils.Tuple
 import Agda.Utils.Lens
 
@@ -156,17 +158,22 @@ coverageCheck f t cs = do
     [ text "generated split tree for" <+> prettyTCM f
     , text $ show splitTree
     ]
-  -- report an error if there are uncovered cases
+  -- report a warning if there are uncovered cases,
+  -- generate a catch-all clause with a metavariable as its body to avoid
+  -- internal errors in the reduction machinery.
   unless (null pss) $
-      setCurrentRange cs $
-        typeError $ CoverageFailure f pss
+      setCurrentRange cs $ do
+        warning $ CoverageIssue f pss
+        let catchAll = Clause noRange noRange empty empty Nothing empty True
+        modifySignature $ updateDefinition f $ updateTheDef
+                        $ updateFunClauses (++ [catchAll])
   -- is = indices of unreachable clauses
   let is = Set.toList $ Set.difference (Set.fromList [0..genericLength cs - 1]) used
   -- report an error if there are unreachable clauses
   unless (null is) $ do
       let unreached = map (cs !!) is
-      setCurrentRange unreached $
-        typeError $ UnreachableClauses f $ map namedClausePats unreached
+      setCurrentRange (map clauseFullRange unreached) $
+        warning $ UnreachableClauses f $ map namedClausePats unreached
   return splitTree
 
 -- | Top-level function for eliminating redundant clauses in the interactive
@@ -184,7 +191,7 @@ type CoverResult = (SplitTree, Set Nat, [(Telescope, [NamedArg DeBruijnPattern])
 cover :: QName -> [Clause] -> SplitClause ->
          TCM CoverResult
 cover f cs sc@(SClause tel ps _ _ target) = do
-  reportSDoc "tc.cover.cover" 10 $ vcat
+  reportSDoc "tc.cover.cover" 10 $ inTopContext $ vcat
     [ text "checking coverage of pattern:"
     , nest 2 $ text "tel  =" <+> prettyTCM tel
     , nest 2 $ text "ps   =" <+> do addContext tel $ prettyTCMPatternList ps
@@ -335,7 +342,8 @@ inferMissingClause f (SClause tel ps _ mpsub (Just t)) = setCurrentRange f $ do
                   Instance  -> newIFSMeta "" (unArg t)
                   Hidden    -> __IMPOSSIBLE__
                   NotHidden -> __IMPOSSIBLE__
-    return $ Clause { clauseRange     = noRange
+    return $ Clause { clauseLHSRange  = noRange
+                    , clauseFullRange = noRange
                     , clauseTel       = tel
                     , namedClausePats = ps
                     , clauseBody      = Just rhs
@@ -502,10 +510,6 @@ computeNeighbourhood delta1 n delta2 d pars ixs hix tel ps mpsub c = do
   con <- liftTCM $ getConForm c
   con <- return $ con { conName = c }  -- What if we restore the current name?
                                        -- Andreas, 2013-11-29 changes nothing!
-{-
-  con <- conSrcCon . theDef <$> getConstInfo con
-  Con con ci [] <- liftTCM $ ignoreSharing <$> (constructorForm =<< normalise (Con con ci []))
--}
 
   -- Get the type of the constructor
   ctype <- liftTCM $ defType <$> getConInfo con
@@ -517,12 +521,26 @@ computeNeighbourhood delta1 n delta2 d pars ixs hix tel ps mpsub c = do
         Just cixs = allApplyElims es
     return (gamma0, cixs)
 
+  -- Andreas, 2017-01-21, issue #2424
+  -- When generating new variable names for the split,
+  -- respect the user written names!
+  let maybeUserWritten arg | getOrigin arg == UserWritten = Just $ unArg arg
+                           | otherwise = Nothing
+      (userNames0 :: [ArgName]) = mapMaybe maybeUserWritten $ teleArgNames tel
+      (userNames :: [PatVarName]) = map dbPatVarName $ lefts $
+        mapMaybe maybeUserWritten $ patternVars ps
+      -- The name of the variable we split is of course reusable!
+      avoidNames = userNames List.\\ [n, "_", "()"]
+      avoidUserName :: ArgName -> ArgName
+      avoidUserName = nameVariant (`elem` avoidNames)
+  debugNames userNames0 userNames avoidNames
+
   -- Andreas, 2012-02-25 preserve name suggestion for recursive arguments
   -- of constructor
 
   let preserve (x, t@(El _ (Def d' _))) | d == d' = (n, t)
       preserve (x, (El s (Shared p))) = preserve (x, El s $ derefPtr p)
-      preserve p = p
+      preserve (x, t) = (avoidUserName x, t)
       gammal = map (fmap preserve) . telToList $ gamma0
       gamma  = telFromList gammal
       delta1Gamma = delta1 `abstract` gamma
@@ -570,17 +588,23 @@ computeNeighbourhood delta1 n delta2 d pars ixs hix tel ps mpsub c = do
 
       debugTel "delta'" delta'
       debugSubst "rho" rho
-      addContext tel $ debugPs ps
+      debugPs tel ps
 
       -- Apply the substitution
       let ps' = applySubst rho ps
-      addContext delta' $ debugPlugged ps'
+      debugPlugged delta' ps'
 
       let mpsub' = applySubst (fromPatternSubstitution rho) mpsub
 
       return $ Just $ SClause delta' ps' rho mpsub' Nothing -- target fixed later
 
   where
+    debugNames userNames0 userNames avoidNames =
+      liftTCM $ reportSDoc "tc.cover.split.con" 20 $ vcat
+        [ text "  user written names in pat.tel  =" <+> sep (map text userNames0)
+        , text "  user written names in patterns =" <+> sep (map text userNames)
+        , text "  names to be avoided by split   =" <+> sep (map text avoidNames)
+        ]
     debugInit con ctype d pars ixs cixs delta1 delta2 gamma tel ps hix =
       liftTCM $ reportSDoc "tc.cover.split.con" 20 $ vcat
         [ text "computeNeighbourhood"
@@ -588,14 +612,14 @@ computeNeighbourhood delta1 n delta2 d pars ixs hix tel ps mpsub c = do
           [ text "context=" <+> (inTopContext . prettyTCM =<< getContextTelescope)
           , text "con    =" <+> prettyTCM con
           , text "ctype  =" <+> prettyTCM ctype
-          , text "ps     =" <+> do addContext tel $ prettyTCMPatternList ps
+          , text "ps     =" <+> do inTopContext $ addContext tel $ prettyTCMPatternList ps
           , text "d      =" <+> prettyTCM d
-          , text "pars   =" <+> prettyList (map prettyTCM pars)
-          , text "ixs    =" <+> addContext delta1 (prettyList (map prettyTCM ixs))
-          , text "cixs   =" <+> do addContext gamma $ prettyList (map prettyTCM cixs)
-          , text "delta1 =" <+> prettyTCM delta1
-          , text "delta2 =" <+> prettyTCM delta2
-          , text "gamma  =" <+> prettyTCM gamma
+          , text "pars   =" <+> do prettyList $ map prettyTCM pars
+          , text "ixs    =" <+> do addContext delta1 $ prettyList $ map prettyTCM ixs
+          , text "cixs   =" <+> do addContext gamma  $ prettyList $ map prettyTCM cixs
+          , text "delta1 =" <+> do inTopContext $ prettyTCM delta1
+          , text "delta2 =" <+> do inTopContext $ addContext delta1 $ addContext gamma $ prettyTCM delta2
+          , text "gamma  =" <+> do inTopContext $ addContext delta1 $ prettyTCM gamma
           , text "hix    =" <+> text (show hix)
           ]
         ]
@@ -616,15 +640,17 @@ computeNeighbourhood delta1 n delta2 d pars ixs hix tel ps mpsub c = do
         [ text (s ++ " =") <+> prettyTCM tel
         ]
 
-    debugPs ps =
-      liftTCM $ reportSDoc "tc.cover.split.con" 20 $ nest 2 $ vcat
-        [ text "ps     =" <+> prettyTCMPatternList ps
-        ]
+    debugPs tel ps =
+      liftTCM $ reportSDoc "tc.cover.split.con" 20 $
+        inTopContext $ addContext tel $ nest 2 $ vcat
+          [ text "ps     =" <+> prettyTCMPatternList ps
+          ]
 
-    debugPlugged ps' =
-      liftTCM $ reportSDoc "tc.cover.split.con" 20 $ nest 2 $ vcat
-        [ text "ps'    =" <+> do prettyTCMPatternList ps'
-        ]
+    debugPlugged delta' ps' =
+      liftTCM $ reportSDoc "tc.cover.split.con" 20 $
+        inTopContext $ addContext delta' $ nest 2 $ vcat
+          [ text "ps'    =" <+> do prettyTCMPatternList ps'
+          ]
 
 -- | Entry point from @Interaction.MakeCase@.
 splitClauseWithAbsurd :: SplitClause -> Nat -> TCM (Either SplitError (Either SplitClause Covering))
@@ -770,7 +796,7 @@ split' ind fixtarget sc@(SClause tel ps _ mpsub target) (BlockingVar x mcons) = 
     inContextOfDelta2 = addContext tel . escapeContext x
 
     -- Debug printing
-    debugInit tel x ps mpsub = liftTCM $ do
+    debugInit tel x ps mpsub = liftTCM $ inTopContext $ do
       reportSDoc "tc.cover.top" 10 $ vcat
         [ text "TypeChecking.Coverage.split': split"
         , nest 2 $ vcat