+++ src/full/Agda/TypeChecking/Serialise/Instances/Internal.hs
@@ -383,11 +383,11 @@ instance EmbPrj TermHead where
     valu _      = malformed
 
 instance EmbPrj I.Clause where
-  icod_ (Clause a b c d e f) = icode6' a b c d e f
+  icod_ (Clause a b c d e f g) = icode7' a b c d e f g
 
   value = vcase valu where
-    valu [a, b, c, d, e, f] = valu6 Clause a b c d e f
-    valu _                  = malformed
+    valu [a, b, c, d, e, f, g] = valu7 Clause a b c d e f g
+    valu _                     = malformed
 
 instance EmbPrj I.ConPatternInfo where
   icod_ (ConPatternInfo a b) = icode2' a b